[ADD] Implement LoginForm component, use custom hooks for login logic, and update LoginPage to utilize the new form structure
