+++ src/features/auth/schemas/authSchemas.jsx
@@ -1,0 +1,6 @@
+import { z } from "zod"
+
+export const loginSchema = z.object({
+    username: z.string().min(3).max(50).or(z.string().email()),
+    password: z.string().min(6)
+})
